requirements.txt
click==8.1.8
datasets==4.0.0
deepdiff>=8.6.1
docstring-parser==0.16
fastapi==0.116.1
fsspec==2024.12.0
geographiclib==2.0
huggingface-hub==0.33.4
inputimeout==1.0.4
Jinja2==3.1.6
joblib==1.4.2
litellm==1.71.1
mammoth==1.8.0
markdownify==0.14.1
mcp[cli]==1.11.0
mysql-connector-python==9.2.0
numpy==2.2.6
pandas==2.2.3
pathvalidate==3.2.1
pdfminer.six==20231228
phonenumbers==8.13.53
pillow==11.1.0
polars-lts-cpu==1.33.1
puremagic==1.27
pydantic==2.10.6
pypdf==6.0.0
python-dotenv==1.0.1
python-pptx==1.0.2
rapidfuzz==3.12.1
retry==0.9.2
s3fs==2024.12.0
starlette==0.47.2
strawberry-graphql==0.275.5
termcolor==2.5.0
tqdm_joblib==0.0.4
uvicorn==0.35.0
wsproto==1.2.0
